Add agents tabs
